requirements.txt
yacs
efficientnet_pytorch
trains